Add ListNode class inside LinkedList
